+++ homeassistant/components/bluesound/const.py
@@ -6,3 +6,5 @@ SERVICE_CLEAR_TIMER = "clear_sleep_timer"
 SERVICE_JOIN = "join"
 SERVICE_SET_TIMER = "set_sleep_timer"
 SERVICE_UNJOIN = "unjoin"
+ATTR_BLUESOUND_GROUP = "bluesound_group"
+ATTR_MASTER = "master"
